Rename knot thing source files
