+++ src/knot_thing_main.h
